fix compileModuleAndReadAST write to InMemoryOutputFileSystem (#793)
Apparently the correct way to call `compileModule` has changed. Our call to `compileModule` is just a copy of the later call in this function, so I pasted in the new call code and it fixed our internal test.

I also have an open source test for this checked into the swift repo's tensorflow branch, so I'm going to separately investigate why that did not trigger.