ExpressionParser: adjust triple handling for Swift REPL on Linux
The OS version is not normally part of the triple on Linux, and will
unnecessarily hamper the ability to use the Swift standard library as
the kernel version changes out-of-sync with the userspace.  The version
information on Linux is not particularly useful, and the binary
compatibility guarantees of the Linux kernel make it unnecessary.

On Windows, the version information is normally associated to the
MSVC toolchain version, not the OS version.  Furthermore, Windows has
strict guarantees about backwards guarantees, making the test less
useful.

This adjusts the triple computation to limit versioning to Darwin
platforms.