Revert "Revert "Add support for fetching signed values from tagged pointers.""
This reverts commit 602ab188a7e18b97d9af95e17271e8fbee129081.

The patch replicated an lldbassert for a certain type of NSNumber for tagged
pointers.  This really shouldn't be an assert since we don't do anything wrong
with these numbers, we just don't print a summary.  So this patch changed the
lldbassert to a log message in reverting the revert.

(cherry picked from commit be0ced03ba9bfab6fcb1fd2c263a33bc6a359cd8)