X86: enable proper frame lowering for swift async on Win64
The Swift asynchronous calling convention requires some additional setup
on the frame.  In particular, the swift async register needs to be
spilled relative to the frame pointer, which we now accommodate.  This
may require some further refinement due to frame setup restrictions on
Windows, but this initial approach should be sufficient to move towards
a working asynchronous support.