[lldb] Call DestroyImpl for eStateInvalid
The process ends up in eStateInvalid when you attempt to interrupt a
GDBRemoteProcess and it doesn't return in time. When that happens we
really should be calling Destroy because otherwise we might end up in a
situation where we receive a packet while or after we've torn down the
main process.