Fix a bug in OptimizedStructLayout when filling gaps before fixed fields with highly-aligned flexible fields.
The code was not considering the possibility that aligning
the current offset to the alignment of a queue might push
us past the end of the gap.  Subtracting the offsets to
figure out the maximum field size for the gap then overflowed,
making us think that we had nearly unbounded space to fill.

Fixes PR 51131.

(cherry picked from commit 326a5a2658d81db46a78b184fe42e522ef170f32)