[lldb] Leak the realized class list
LLDB JITs a call to objc_copyRealizedClassList in the inferior and frees
the result. If a project defines its own `free` or `Free` function, LLDB
can incorrectly resolve it to the wrong symbol and corrupt the
Objective-C runtime. As an stopgap solution we'll leak the memory until
we can fix the symbol resolution.

rdar://81241350