clang: Add -fcas-token-cache to cache raw tokens
Add -fcas-token-cache, which does a raw-preprocess of each file in
isolation, producing and caching raw tokens, before including it in the
current context.

Note that this was bootstrapped by starting with a partial revert of
0a6b5b653ee47234674614fecc213c1d73bb1e28, which removed the PTH feature,
with the primary differences being:

- CAS-based workflow
- Raw tokens are cached (raw identifiers), allowing them to be reused
  from different macro contexts

This implementation strategy has some flaws, mainly along error paths.
(Also it lacks tests entirely!)

Likely the "right" implementation would layer full tokenization on top
of raw tokenization, so that the two paths (with/without caching)
hit the same logic.

Nevertheless, this gets us some nice speedups when scanning and works
well when there are no preprocessor errors, so it's useful as a proof of
concept.