Run coro::salavageDebugInfo on dbg.declares in entry function
The patch also calls coro::salavageDebugInfo() on all dbg.declares in
entry function. This is necessary for Swift because in the Swift
calling convention the entry function can already have a coroutine
context argument. I believe that the patch is neutral for C++
coroutines at -O0 (and mildly beneficial at -O), as it does rewrite
dbg.declares to point into the freshly setup coroutine frame, as shown
in the updated test.

rdar://82037718