cas.o: Remove bloat from deadstrip=compile symbol table
Symbols should only be added to the list of
IndirectDeadStripCompileSymbols if they are deadstrip=compile and they
are indirectly referenced (by name). The logic in
`CompileUnitBuilder::getOrCreateTarget()` was incorrectly adding all
symbols that were indirectly referenced after being defined. Fix that,
aligning with the logic in `createSymbol()`.

For `--prefer-indirect-symbol-refs` this cuts out the number of children
from symbol tables by 20%, saves 2% total cost, when ingesting
lib/Support no-opt object files. I confirmed the number of symbols stays
the same so nothing was "lost".

Sadly I'm going to be lazy and add a FIXME for the test instead of
writing one...